// next/types/types.ts
export interface Category {
  name: string
}

export interface Image {
  url: string
  alternativeText: string
}

export interface Author {
  name: string
  avatar: Image
  slug: string
}

export interface Article {
  title: string
  description: string
  slug: string
  content: string
  dynamic_zone: any[]
  createdAt: string
  updatedAt: string
  publishedAt: string
  locale: string
  image: Image
  categories: Category[]
  author: Author
}

export interface Product {
  id: number
  name: string
  slug: string
  description: string
  price: number
  plans: any[]
  perks: any[]
  featured?: boolean
  images: any[]
  categories?: any[]
}